+++ frontend/core-ui/src/modules/contacts/companies/components/companies-command-bar/CompaniesCommandBar.tsx
@@ -1,8 +1,7 @@
 import { ApolloError } from '@apollo/client';
 import { CommandBar, RecordTable, Separator, toast } from 'erxes-ui';
-import { Export, TagsSelect } from 'ui-modules';
+import { Can, Export, TagsSelect } from 'ui-modules';
 import { CompaniesDelete } from './CompaniesDelete';
-import { CompaniesMerge } from './CompaniesMerge';
 
 export const CompaniesCommandBar = () => {
   const { table } = RecordTable.useRecordTable();
@@ -55,7 +54,7 @@ export const CompaniesCommandBar = () => {
             },
           })}
         />
-        <Separator.Inline/>
+        <Separator.Inline />
         <Export
           pluginName="core"
           moduleName="contact"
@@ -63,7 +62,6 @@ export const CompaniesCommandBar = () => {
           buttonVariant="secondary"
           ids={companyIds}
         />
-        <Separator.Inline />
         {/* <CompaniesMerge
           companies={table
             .getFilteredSelectedRowModel()
@@ -72,7 +70,12 @@ export const CompaniesCommandBar = () => {
           rows={selectedRows} 
         /> */}
         {/* <Separator.Inline /> */}
-        <CompaniesDelete companyIds={companyIds} rows={selectedRows} />
+        <Can action="contactsDelete">
+          <>
+            <Separator.Inline />
+            <CompaniesDelete companyIds={companyIds} rows={selectedRows} />
+          </>
+        </Can>
       </CommandBar.Bar>
     </CommandBar>
   );