+++ frontend/plugins/tourism_ui/src/modules/tms/branch-detail/dashboard/category/graphql/queries.ts
@@ -1,8 +1,8 @@
 import { gql } from '@apollo/client';
 
 export const GET_CATEGORIES = gql`
-  query BmsTourCategories($parentId: String) {
-    bmsTourCategories(parentId: $parentId) {
+  query BmsTourCategories($parentId: String, $name: String) {
+    bmsTourCategories(parentId: $parentId, name: $name) {
       _id
       name
       code