+++ frontend/plugins/tourism_ui/src/modules/tms/branch-detail/dashboard/itinerary/graphql/mutation.ts
@@ -7,13 +7,14 @@ export const CREATE_ITINERARY = gql`
     $branchId: String
     $name: String
     $duration: Int
-    $color: String
+    $images: [String]
     $guideCost: Float
     $driverCost: Float
     $foodCost: Float
     $gasCost: Float
     $personCost: JSON
     $guideCostExtra: Float
+    $content: String
   ) {
     bmsItineraryAdd(
       groupDays: $groupDays
@@ -21,13 +22,14 @@ export const CREATE_ITINERARY = gql`
       branchId: $branchId
       name: $name
       duration: $duration
-      color: $color
+      images: $images
       guideCost: $guideCost
       driverCost: $driverCost
       foodCost: $foodCost
       gasCost: $gasCost
       personCost: $personCost
       guideCostExtra: $guideCostExtra
+      content: $content
     ) {
       _id
     }
@@ -42,13 +44,14 @@ export const EDIT_ITINERARY = gql`
     $duration: Int
     $totalCost: Float
     $groupDays: [DayItemInput]
-    $color: String
+    $images: [String]
     $foodCost: Float
     $personCost: JSON
     $gasCost: Float
     $driverCost: Float
     $guideCost: Float
     $guideCostExtra: Float
+    $content: String
   ) {
     bmsItineraryEdit(
       _id: $id
@@ -57,13 +60,14 @@ export const EDIT_ITINERARY = gql`
       duration: $duration
       totalCost: $totalCost
       groupDays: $groupDays
-      color: $color
+      images: $images
       foodCost: $foodCost
       personCost: $personCost
       gasCost: $gasCost
       driverCost: $driverCost
       guideCost: $guideCost
       guideCostExtra: $guideCostExtra
+      content: $content
     ) {
       _id
     }